+++ tests/testthat/test-get_model_designations.R
@@ -2,7 +2,7 @@ context("get_model_designations")
 library(covidHubUtils)
 library(dplyr)
 
-test_that("get_model_designations works: local directory, all models", {
+test_that("get_model_designations works: local directory, data-processed folder, all models", {
   actual <- covidHubUtils::get_model_designations(
     source = "local_hub_repo",
     hub_repo_path = "test-data/test-get_model_designations"
@@ -16,13 +16,30 @@ test_that("get_model_designations works: local directory, all models", {
     designation = c(
       "proposed", "primary", "primary", "secondary", "proposed",
       "other", "other"
-    )
+    ),
+    stringsAsFactors = FALSE
   )
 
   expect_equal(actual, expected)
 })
 
-test_that("get_model_designations works: local directory, specified models", {
+test_that("get_model_designations works: local directory, data-forecasts folder, all models", {
+  actual <- covidHubUtils::get_model_designations(
+    source = "local_hub_repo",
+    hub_repo_path = "test-data/test-get_model_designations",
+    hub = "FluSight"
+  )
+  
+  expected <- data.frame(
+    model = c("TeamA-ModelA","TeamB-ModelB"),
+    designation = c("primary", "primary"),
+    stringsAsFactors = FALSE
+  )
+  
+  expect_equal(actual, expected)
+})
+
+test_that("get_model_designations works: local directory, data-processed folder, specified models", {
   actual <- covidHubUtils::get_model_designations(
     source = "local_hub_repo",
     models = c("COVIDhub-baseline", "COVIDhub-ensemble", "teamA-modelA"),
@@ -31,9 +48,27 @@ test_that("get_model_designations works: local directory, specified models", {
 
   expected <- data.frame(
     model = c("COVIDhub-baseline", "COVIDhub-ensemble", "teamA-modelA"),
-    designation = c("proposed", "primary", "primary")
+    designation = c("proposed", "primary", "primary"),
+    stringsAsFactors = FALSE
   )
 
+  expect_equal(actual, expected)
+})
+
+test_that("get_model_designations works: local directory, data-forecasts folder, specified models", {
+  actual <- covidHubUtils::get_model_designations(
+    source = "local_hub_repo",
+    models = c("TeamA-ModelA"),
+    hub_repo_path = "test-data/test-get_model_designations",
+    hub = "FluSight"
+  )
+  
+  expected <- data.frame(
+    model = c("TeamA-ModelA"),
+    designation = c("primary"),
+    stringsAsFactors = FALSE
+  )
+  
   expect_equal(actual, expected)
 })
 
@@ -46,7 +81,8 @@ test_that("get_model_designations works: zoltar, specified models", {
 
   expected <- data.frame(
     model = c("COVIDhub-baseline", "COVIDhub-ensemble"),
-    designation = c("secondary", "primary")
+    designation = c("secondary", "primary"),
+    stringsAsFactors = FALSE
   )
 
   expect_true(dplyr::all_equal(actual, expected))
@@ -61,7 +97,8 @@ test_that("get_model_designations work: local hub repo, space in hub repo path",
 
   expected <- data.frame(
     model = c("teamA-modelA", "teamB-modelB"),
-    designation = c("primary", "secondary")
+    designation = c("primary", "secondary"),
+    stringsAsFactors = FALSE
   )
 
   expect_true(dplyr::all_equal(actual, expected))