+++ tests/testthat/test-get_model_metadata.R
@@ -2,7 +2,7 @@ context("get_model_metadata")
 library(covidHubUtils)
 library(dplyr)
 
-test_that("get_model_metadata works: local directory, all models", {
+test_that("get_model_metadata works: local directory, data-processed folder, all models", {
   actual <- covidHubUtils::get_model_metadata(
     source = "local_hub_repo",
     hub_repo_path = "test-data/test-get_model_designations"
@@ -24,13 +24,36 @@ test_that("get_model_metadata works: local directory, all models", {
     ),
     ensemble_of_hub_models = c(
       NA, TRUE, NA, NA, NA, NA, NA
-      )
+    ),
+    stringsAsFactors = FALSE
   )
 
   expect_equal(actual, expected)
 })
 
-test_that("get_model_metadata works: local directory, specified models", {
+test_that("get_model_metadata works: local directory, data-forecasts folder, all models", {
+  actual <- covidHubUtils::get_model_metadata(
+    source = "local_hub_repo",
+    hub_repo_path = "test-data/test-get_model_designations",
+    hub = c("FluSight")
+  )
+  
+  expected <- data.frame(
+    team_name = c("TeamA", "TeamB"),
+    model_name = c("ModelA", "ModelB"),
+    model = c("TeamA-ModelA","TeamB-ModelB"),
+    website_url = c(
+      "https://covid19forecasthub.org/", "https://covid19forecasthub.org/"
+    ),
+    designation = c("primary", "primary"),
+    ensemble_of_hub_models = c(FALSE, FALSE),
+    stringsAsFactors = FALSE
+  )
+  
+  expect_equal(actual, expected)
+})
+
+test_that("get_model_metadata works: local directory, data-processed folder, specified models", {
   actual <- covidHubUtils::get_model_metadata(
     source = "local_hub_repo",
     models = c("COVIDhub-baseline", "COVIDhub-ensemble", "teamA-modelA"),
@@ -43,9 +66,31 @@ test_that("get_model_metadata works: local directory, specified models", {
     model = c("COVIDhub-baseline", "COVIDhub-ensemble", "teamA-modelA"),
     designation = c("proposed", "primary", "primary"),
     website_url = c(NA, "https://covid19forecasthub.org/", NA),
-    ensemble_of_hub_models = c(NA, TRUE, NA)
+    ensemble_of_hub_models = c(NA, TRUE, NA),
+    stringsAsFactors = FALSE
   )
 
+  expect_equal(actual, expected)
+})
+
+test_that("get_model_metadata works: local directory, data-forecasts folder, specified models", {
+  actual <- covidHubUtils::get_model_metadata(
+    source = "local_hub_repo",
+    models = c("TeamA-ModelA"),
+    hub_repo_path = "test-data/test-get_model_designations",
+    hub = c("FluSight")
+  )
+  
+  expected <- data.frame(
+    team_name = c("TeamA"),
+    model_name = c("ModelA"),
+    model = c("TeamA-ModelA"),
+    website_url = c("https://covid19forecasthub.org/"),
+    designation = c("primary"),
+    ensemble_of_hub_models = c(FALSE),
+    stringsAsFactors = FALSE
+  )
+  
   expect_equal(actual, expected)
 })
 
@@ -60,7 +105,8 @@ test_that("get_model_metadata works: local hub repo, space in hub repo path", {
     team_name = c(rep("test team name", 2)),
     model_name = c("A", "teamB"),
     model = c("teamA-modelA", "teamB-modelB"),
-    designation = c("primary", "secondary")
+    designation = c("primary", "secondary"),
+    stringsAsFactors = FALSE
   )
 
   expect_true(dplyr::all_equal(actual, expected))