+++ tests/testthat/test-load_truth.R
@@ -60,6 +60,19 @@ test_that("default selections from remote hub repo", {
       expected_ecdc_inc_hosp
     )
   ))
+  
+  # Flu hub
+  actual_flu <- load_truth(hub = "FluSight")
+  # weekly
+  expected_flu_inc_flu_hosp <- load_truth(
+    truth_source = c("HealthData"),
+    target_variable = c("inc flu hosp"),
+    hub = c("FluSight")
+  )
+  
+  expect_true(dplyr::all_equal(
+    actual_flu,
+    expected_flu_inc_flu_hosp))
 })
 
 test_that("comapre target variables and models on default selections from remote hub 
@@ -114,19 +127,17 @@ test_that("load one target variable from multiple sources from remote hub repo",
     )
   )
   # ECDC hub
-  actual_ecdc <- load_truth(
-    truth_source = c("JHU", "ECDC"),
-    target_variable = c("inc case"),
-    hub = c("ECDC")
+  # no test cases
+  
+  # Flu hub
+  actual_flu <- load_truth(
+    truth_source = c("HealthData"),
+    target_variable = c("inc flu hosp"),
+    hub = c("FluSight")
   )
-  expect_equal(unique(actual_ecdc$target_variable), c("inc case"))
-  expect_equal(
-    unique(actual_ecdc$model),
-    c(
-      "Observed Data (ECDC)",
-      "Observed Data (JHU)"
-    )
-  )
+  
+  expect_equal(unique(actual_flu$target_variable), c("inc flu hosp"))
+  expect_equal(unique(actual_flu$model), c("Observed Data (HealthData)"))
 })
 
 test_that("handles `inc hosp` and `HealthData` source in US hub correctly when loading
@@ -173,35 +184,12 @@ test_that("handles `inc hosp` and `HealthData` source in US hub correctly when l
 test_that("handles `ECDC`source in ECDC hub correctly when loading from remote 
           hub repo", {
   # for case/death data, return warning and weekly data when temporal resolution is daily
-  expect_warning(actual <- load_truth(
+  expect_error(actual <- load_truth(
     truth_source = c("ECDC"),
     target_variable = c("inc case", "inc death"),
     temporal_resolution = "daily",
     hub = c("ECDC")
   ))
-
-  expect_equal(
-    unique(actual$target_end_date),
-    seq(
-      min(actual$target_end_date),
-      max(actual$target_end_date), 7
-    )
-  )
-  expect_equal(
-    unique(weekdays(actual$target_end_date)),
-    "Saturday"
-  )
-  expect_equal(
-    unique(weekdays(actual$week_start)),
-    "Monday"
-  )
-  expect_equal(
-    unique(actual$target_end_date),
-    seq(
-      min(actual$target_end_date),
-      max(actual$target_end_date), 7
-    )
-  )
 })
 
 test_that("expects warnings when loading versioned data from hub repo", {
@@ -212,3 +200,13 @@ test_that("expects warnings when loading versioned data from hub repo", {
     as_of = "2020-11-23"
   ))
 })
+
+test_that("expects error when loading flu hub data from covidData", {
+  expect_error(data <- load_truth(
+    hub = c("FluSight"),
+    data_location = "covidData"
+  ))
+})
+
+
+